creat with doubleTotal

+++ database/migrations/2024_04_01_165404_create_projects_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('projects', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->float('Total');
+            $table->double('Total');
             $table->text('description');
             $table->date('start_date');
             $table->date('end_date');
